refactor: domain, util 패키지 그룹
