+++ src/main/java/step4/carracing/util/InputView.java
@@ -1,4 +1,4 @@
-package step4.carracing;
+package step4.carracing.util;
 
 import java.util.Arrays;
 import java.util.List;
