+++ src/main/java/step4/carracing/util/ResultView.java
@@ -1,4 +1,4 @@
-package step4.carracing;
+package step4.carracing.util;
 
 import step4.carracing.domain.Car;
 import step4.carracing.domain.Cars;
