+++ src/test/java/step4/carracing/util/CarFactoryTest.java
@@ -1,8 +1,9 @@
-package step4.carracing;
+package step4.carracing.util;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step4.carracing.RandomCarMoveStrategy;
 import step4.carracing.domain.Car;
 import step4.carracing.domain.Cars;
